[llvm-c][Disassembler] Add option to print branch immediate as address
When printing disassembled branching instructions, llvm-objdump
translates the immediate operand into a PC-relative address, making it
easier to understand where the call or jump points to in the list of
instructions [0]. However, this feature was not available for llvm-c's
disassembler. This commit adds a new option for the disassembler, namely
LLVMDisassembler_Option_PrintBranchImmAsAddress, to tell it to use
PC-relative addresses instead of byte offsets.

Additionnally, in the case of x86, we must adjust the resulting address
to account for a size of the current instruction, given that the
immediate operand is relative to the start of the next instruction. This
is equivalent to what llvm-objdump does [1].

Here is the output for an x86_64 program disassembled with llvm-c:

```
// Before
  2b:   jne     2

// After, with LLVMDisassembler_Option_PrintBranchImmAsAddress
  2b:   jne     0x2f

// Disassembling the same code, with libbfd + opcodes
  2b:   jne    0x000000000000002f
```

[0] 5fad05e80dd0 ("[MCInstPrinter] Pass `Address` parameter to MCOI::OPERAND_PCREL typed operands. NFC")
[1] 87de9a0786d9 ("[X86InstPrinter] Change printPCRelImm to print the target address in hexadecimal form")